code/requirements.txt
pybids==0.14.0
jupyterlab
black
nb_black
datalad
rich
pytest
